Change how end-of-model effects are treated in Obj. Function
This commit makes a small change to the objective function -- particularly
around how loan costs occuring past the final model period are accounted for.

Previously, the total loan costs were multiplied by the fraction of (number of
years left in the model) over (lifetime of the technology) or 1.0 -- whichever was
smaller. This essentially removed the loan payments that occured past the model's
time horizon.

This has been reformulated so that the denominator of the fraction is the loan lifetime
(or capital recovery period). This is because loans are paid over the loan lifetime, not
the technical lifetime.

Furthermore, we have made a slight edit to how this is encoded so that it works for a
GlobalDiscountRate of 0. Previously, GDRs of 0 caused an error.