Add constraints for technology groups
This commit adds to the TEMOA formulation new constraints for technology groups, specifically: MinInputGroup, MaxInputGroup, MinOutputGroup, MaxOutputGroup. Such constraints allow the modeler to specify  the minimum/maximum share of a specific commodity with respect to the total input/output commodities flow for a technology group. For instance, MaxInputGroup can be used to set the maximum percentage of natural gas with respect to the total input of technologies for electricity production. The technology group can be built with technologies included in the "tech_groups" set, group names in "groups" and by associating technologies to their group through the "TechGroupWeight" table in "temoa_schema.sql".