Update Storage Equations (no seasonal storage)
This commit revamps how storage technologies are treated.
Previously, storage technologies could do both intra-season
and inter-season storage... So, in a representative week setting
the storage techs could do inter-week storage. This is not wanted
since Temoa naively assumes the representative periods occur in a
weighted succession.

This commit fixes that by enforcing that the state-of-charge of
storage technologies at the first and final timestep of each
representative period (i.e. 'season' in Temoa-speak) are equal.